# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User System API",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/datasets": {
            "get": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Dataset show",
                "operationId": "03a25edb34565c70dbac2d4cbb96f1fd",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Dataset"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Dataset update",
                "operationId": "c8720acc01283a7c6f0e3bf1e2591e15",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Dataset"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dataset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/datasets/{id}": {
            "get": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Dataset get by ID",
                "operationId": "9a82b19cc4e22b94a7da5536ccb91813",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dataset id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dataset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "users/{id}/datasets": {
            "get": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Get datasets of user",
                "operationId": "2908beaa6f5069295b52e6ac50a718c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Dataset"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dataset"
                ],
                "summary": "Create dataset for user",
                "operationId": "b681c12fcdbb297e847154f25e46bdd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Dataset"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dataset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/users/{id}/roles": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add role to the User",
                "operationId": "addrole",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Role type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "406": {
                        "description": "Something went wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete role from the User",
                "operationId": "deleterole",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Role type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "406": {
                        "description": "Something went wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of users",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "operationId": "update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "operationId": "create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by ID",
                "operationId": "get",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/users/attempt": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Attempts user credentials",
                "operationId": "attempt",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "hashed password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Dataset": {
                "title": "Dataset",
                "description": "Dataset model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "example": 1
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "user_id",
                        "example": 1
                    },
                    "type": {
                        "title": "type",
                        "description": "type",
                        "example": "CREATOR"
                    },
                    "content": {
                        "title": "content",
                        "description": "content",
                        "example": "Documents 123"
                    },
                    "status": {
                        "title": "status",
                        "description": "status",
                        "example": "PENDING"
                    },
                    "comment": {
                        "title": "comment",
                        "description": "comment",
                        "example": "Wrong name"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created_at",
                        "example": "2022-01-22T21:34:30.000000"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "example": "2022-01-22T21:34:30.000000"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Dataset"
                }
            },
            "Role": {
                "title": "Role",
                "description": "Role model",
                "properties": {
                    "type": {
                        "title": "type",
                        "description": "APPLICANT or CREATOR or ADMIN or MODERATOR",
                        "example": "APPLICANT"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created_at",
                        "example": "2022-01-22T21:34:30.000000"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "example": "2022-01-22T21:34:30.000000"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Role"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "example": "Ivan"
                    },
                    "surname": {
                        "title": "surname",
                        "description": "surname",
                        "example": "Hue"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "example": "vaehu@gmail.com"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "example": "Developer in Google working around 20 hour a week"
                    },
                    "media_id": {
                        "title": "media_id",
                        "description": "media_id",
                        "example": "1"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created_at",
                        "example": "2022-01-22T21:34:30.000000"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "example": "2022-01-22T21:34:30.000000"
                    },
                    "roles": {
                        "title": "roles",
                        "description": "roles",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "API Endpoints of User"
        },
        {
            "name": "Dataset",
            "description": "API Endpoints of Dataset"
        }
    ]
}